refactor(request): extract comment command helpers
